+++ src/pyordgraph.rs
@@ -49,7 +49,7 @@ impl PyOrdGraph {
         Ok(self.G.adjacent(&u, &v))
     }
 
-    pub fn degree(&self, u:Vertex) -> PyResult<usize> {
+    pub fn degree(&self, u:Vertex) -> PyResult<u32> {
         Ok(self.G.degree(&u))
     }
